+++ src/components/List/ListData.tsx
@@ -1,83 +1,32 @@
-import React, { useEffect, useState } from "react";
-import Loading from "../loading";
-import { useUser } from "../../utils/UserProvider";
-import useThrottle from "../../hooks/useThrottle";
-import useListData, { IRepoItem } from "../../hooks/useListData";
-import useSelectItem from "../../hooks/useSelectItem";
-import ListUI from "./ListUI";
+import React from "react";
 
-const List = () => {
-  // get user details
-  const { user, token: auth, isLoggingin } = useUser();
-  // get data details
-  const { data, isLoading, isDeleting, fetchData, deleteSelectedData } =
-    useListData(auth, user);
-  // get selected item details
-  const { selectedItems, toggleSelectAll, toggleSelect, clearSelection } =
-    useSelectItem<IRepoItem>();
-  // validation
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isValid, setIsValid] = useState(false);
+import ListItem from "./ListItem";
 
+const List = ({ items, selectedItems, isDeleting, toggleSelect }) => {
   const handleToggleSelect = (item: IRepoItem) => {
     toggleSelect(item);
   };
 
-  const handleSelectAll = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.checked) {
-      // Select all repositories
-      toggleSelectAll(data);
-    } else {
-      // Deselect all repositories
-      clearSelection();
-    }
-  };
-
-  const handleDelete = async () => {
-    setIsModalOpen(false);
-    deleteSelectedData(selectedItems);
-    clearSelection();
-  };
-
-  const throttleRefresh = useThrottle(async () => {
-    fetchData();
-  }, 3000);
-
-  const handleValidation = (value: string) => {
-    // validate input matches "{owner}" using regex
-    const regex = new RegExp(user, "i");
-    if (!regex.test(value)) {
-      return setIsValid(false);
-    }
-    return setIsValid(true);
-  };
-
-  useEffect(() => {
-    if (user) throttleRefresh();
-  }, [user, isLoggingin]);
-
-  if (!user && !isLoggingin) {
-    return;
-  }
   return (
-    <Loading isLoading={isLoading || isLoggingin}>
-      <ListUI
-        {...{
-          user,
-          data,
-          handleSelectAll,
-          throttleRefresh,
-          selectedItems,
-          handleToggleSelect,
-          setIsModalOpen,
-          isDeleting,
-          isModalOpen,
-          isValid,
-          handleDelete,
-          handleValidation,
-        }}
-      />
-    </Loading>
+    <>
+      <div className="max-h-screen pr-4 overflow-auto ">
+        {items.map((repo, i) => (
+          <ListItem
+            key={i}
+            checked={selectedItems?.some((item) => item.value === repo.value)}
+            isLoading={
+              selectedItems?.find(({ value }) => value === repo.value) &&
+              isDeleting
+            }
+            id={i}
+            value={repo.value}
+            onChange={({ target }) =>
+              handleToggleSelect({ id: i, value: target.value })
+            }
+          />
+        ))}
+      </div>
+    </>
   );
 };
